Break indexing out into its own module
