+++ src/toil_vg/vg_evaluation_pipeline.py
@@ -27,7 +27,9 @@ from toil.common import Toil
 from toil.job import Job
 from toil_lib.toillib import *
 from toil_lib.programs import docker_call
-from toil_vg.dockered_chunked_call import *
+from toil_vg.vg_common import *
+from toil_vg.vg_call import *
+from toil_vg.vg_index import *
 
 def parse_args():
     """
@@ -69,32 +71,20 @@ def parse_args():
         help="Name of reference path in the graph (eg. ref or 17)")
     parser.add_argument("--path_size", nargs='+', type=int,
         help="Size of the reference path in the graph")    
-    parser.add_argument("--edge_max", type=int, default=5,
-        help="maximum edges to cross in index")
-    parser.add_argument("--kmer_size", type=int, default=10,
-        help="size of kmers to use in indexing and mapping")
     parser.add_argument("--num_fastq_chunks", type=int, default=3,
         help="number of chunks to split the input fastq file records")
     parser.add_argument("--restat", default=False, action="store_true",
         help="recompute and overwrite existing stats files")
-    parser.add_argument("--reindex", default=False, action="store_true",
-        help="don't re-use existing indexed graphs")
-    parser.add_argument("--index_mode", choices=["rocksdb", "gcsa-kmer",
-        "gcsa-mem"], default="gcsa-mem",
-        help="type of vg index to use for mapping")
-    parser.add_argument("--include_pruned", action="store_true",
-        help="use the pruned graph in the index")
-    parser.add_argument("--include_primary", action="store_true",
-        help="use the primary path in the index")
-    parser.add_argument("--index_cores", type=int, default=3,
-        help="number of threads during the indexing step")
     parser.add_argument("--alignment_cores", type=int, default=3,
         help="number of threads during the alignment step")
 
-    # Add common calling options shared with dockered_chunked_call
+    # Add common indexing options shared with vg_index
+    index_parse_args(parser)
+    
+    # Add common calling options shared with vg_call
     chunked_call_parse_args(parser)
 
-    # Add common docker options shared with dockered_chunked_call
+    # Add common docker options
     add_docker_tool_parse_args(parser)
 
     options = parser.parse_args()
@@ -104,217 +94,16 @@ def parse_args():
 
     return parser.parse_args()
 
-def get_files_by_file_size(dirname, reverse=False):
-    """ Return list of file paths in directory sorted by file size """
-
-    # Get list of files
-    filepaths = []
-    for basename in os.listdir(dirname):
-        filename = os.path.join(dirname, basename)
-        if os.path.isfile(filename):
-            filepaths.append(filename)
-
-    # Re-populate list with filename, size tuples
-    for i in xrange(len(filepaths)):
-        filepaths[i] = (filepaths[i], os.path.getsize(filepaths[i]))
-
-    return filepaths
-
-def batch_iterator(iterator, batch_size):
-    """Returns lists of length batch_size.
-
-    This can be used on any iterator, for example to batch up
-    SeqRecord objects from Bio.SeqIO.parse(...), or to batch
-    Alignment objects from Bio.AlignIO.parse(...), or simply
-    lines from a file handle.
-
-    This is a generator function, and it returns lists of the
-    entries from the supplied iterator.  Each list will have
-    batch_size entries, although the final list may be shorter.
-    
-    From http://biopython.org/wiki/Split_large_file
-    """
-    entry = True  # Make sure we loop once
-    while entry:
-        batch = []
-        while len(batch) < batch_size:
-            try:
-                entry = iterator.next()
-            except StopIteration:
-                entry = None
-            if entry is None:
-                # End of file
-                break
-            batch.append(entry)
-        if batch:
-            yield batch
-
 def run_indexing(job, options):
     """
     For each server listed in the server_list tsv, kick off child jobs to
     align and evaluate it.
-
     """
+
+    # common indexing logic, shared between here and standalone, but
+    # both requiring different children. 
+    index_key, index_dir_id = build_indexes(job, options)
     
-    RealTimeLogger.get().info("Starting indexing...")
-    
-    # Set up the IO stores each time, since we can't unpickle them on Azure for
-    # some reason.
-    input_store = IOStore.get(options.input_store)
-    out_store = IOStore.get(options.out_store)
-
-    graph_file = os.path.basename(options.vg_graph)
-
-    # Define work directory for docker calls
-    work_dir = job.fileStore.getLocalTempDir()
-    
-    # Download local input files from the remote storage container
-    graph_dir = work_dir
-    robust_makedirs(graph_dir)
-    
-    graph_filename = "{}/graph.vg".format(graph_dir)
-    graph_file_remote_path = graph_file
-    input_store.read_input_file(graph_file_remote_path, graph_filename)
-    
-    
-    # Now run the indexer.
-    RealTimeLogger.get().info("Indexing {}".format(options.vg_graph))
-            
-    if options.index_mode == "rocksdb":
-        # Make the RocksDB index
-        command = ['vg', 'index', '-s', '-k', str(options.kmer_size), '-e', str(options.edge_max), '-t', str(job.cores), os.path.basename(graph_filename), os.path.basename('{}/{}.index'.format(graph_dir, graph_file))]
-        options.drunner.call(command, work_dir=work_dir)
-
-    elif (options.index_mode == "gcsa-kmer" or
-        options.index_mode == "gcsa-mem"):
-        # We want a GCSA2/xg index. We have to prune the graph ourselves.
-        # See <https://github.com/vgteam/vg/issues/286>.
-
-        # What will we use as our temp combined graph file (containing only
-        # the bits of the graph we want to index, used for deduplication)?
-        to_index_filename = "{}/to_index.vg".format(
-            work_dir)
-
-        # Where will we save the kmers?
-        kmers_filename = "{}/index.graph".format(
-            work_dir)
-
-        
-        with open(to_index_filename, "w") as to_index_file:
-
-            if options.include_pruned:
-
-                RealTimeLogger.get().info("Pruning {} to {}".format(
-                    graph_filename, to_index_filename))
-
-                # Prune out hard bits of the graph
-                # and complex regions
-                # and short disconnected chunks
-                command = [['vg', 'mod', '-p', '-l', str(options.kmer_size),
-                            '-t', str(job.cores), '-e', str(options.edge_max),
-                            os.path.basename(graph_filename)]]
-                command.append(['vg', 'mod', '-S', '-l', str(options.kmer_size * 2),
-                                '-t', str(job.cores), '-'])
-                options.drunner.call(command, work_dir=work_dir, outfile=to_index_file)
-
-            if options.include_primary:
-
-                # Then append in the primary path. Since we don't knoiw what
-                # "it's called, we retain "ref" and all the 19", "6", etc paths
-                # "from 1KG.
-
-                RealTimeLogger.get().info(
-                    "Adding primary path to {}".format(to_index_filename))
-            
-                RealTimeLogger.get().info(
-                    "to_index_file: {}".format(to_index_file))
-                # See
-                # https://github.com/vgteam/vg/issues/318#issuecomment-215102199
-
-                # Generate all the paths names we might have for primary paths.
-                # It should be "ref" but some graphs don't listen
-                RealTimeLogger.get().info("OPTIONS.PATH_NAME: {}".format(options.path_name[0]))
-                RealTimeLogger.get().info("CHR IN OPTIONS.PATH_NAME: {}".format('chr' in options.path_name[0]))
-                if 'chr' in options.path_name[0]:
-                    ref_names = (["ref", "chrx", "chrX", "chry", "chrY", "chrm", "chrM"] +
-                        ['chr'+str(x) for x in xrange(1, 23)])
-                else:
-                    ref_names = (["ref", "x", "X", "y", "Y", "m", "M"] +
-                        [str(x) for x in xrange(1, 23)])
-                RealTimeLogger.get().info("REF_NAMES: {}".format(ref_names))
-                ref_options = []
-                for name in ref_names:
-                    # Put each in a -r option to retain the path
-                    ref_options.append("-r")
-                    ref_options.append(name)
-
-                # Retain only the specified paths (only one should really exist)
-                command = ['vg', 'mod', '-N'] + ref_options + ['-t', str(job.cores), os.path.basename(graph_filename)]
-                options.drunner.call(command, work_dir=work_dir, inputs=[graph_filename],
-                                     outfile=to_index_file)
-                
-        time.sleep(1)
-
-        # Now we have the combined to-index graph in one vg file. We'll load
-        # it (which deduplicates nodes/edges) and then find kmers.
-        RealTimeLogger.get().info("Finding kmers in {} to {}".format(
-            to_index_filename, kmers_filename))
-
-        # Make the GCSA2 kmers file
-        with open(kmers_filename, "w") as kmers_file:
-            command = [['vg', 'view', '-v', os.path.basename(to_index_filename)]]
-            command.append(['vg', 'kmers', '-g', '-B', '-k', str(options.kmer_size),
-                            '-H', '1000000000', '-T', '1000000001', '-t', str(job.cores), '-'])
-            options.drunner.call(command, work_dir=work_dir, outfile=kmers_file)
-
-        time.sleep(1)
-
-        # Where do we put the GCSA2 index?
-        gcsa_filename = graph_filename + ".gcsa"
-
-        RealTimeLogger.get().info("GCSA-indexing {} to {}".format(
-                kmers_filename, gcsa_filename))
-
-        # Make the gcsa2 index. Make sure to use 3 doubling steps to work
-        # around <https://github.com/vgteam/vg/issues/301>
-        command = ['vg', 'index', '-t', str(job.cores), '-i', 
-            os.path.basename(kmers_filename), '-g', os.path.basename(gcsa_filename),
-            '-X', '3', '-Z', '2000']
-        options.drunner.call(command, work_dir=work_dir)
-
-        # Where do we put the XG index?
-        xg_filename = graph_filename + ".xg"
-
-        RealTimeLogger.get().info("XG-indexing {} to {}".format(
-                graph_filename, xg_filename))
-        
-        command = ['vg', 'index', '-t', str(job.cores), '-x', 
-            os.path.basename(xg_filename), os.path.basename(graph_filename)]
-        options.drunner.call(command, work_dir=work_dir)
-
-    else:
-        raise RuntimeError("Invalid indexing mode: " + options.index_mode)
-
-    # Define a file to keep the compressed index in, so we can send it to
-    # the output store.
-    index_dir_tgz = "{}/index.tar.gz".format(
-        job.fileStore.getLocalTempDir())
-
-    # Now save the indexed graph directory to the file store. It can be
-    # cleaned up since only our children use it.
-    RealTimeLogger.get().info("Compressing index of {}".format(
-        graph_filename))
-    index_dir_id = write_global_directory(job.fileStore, graph_dir,
-        cleanup=True, tee=index_dir_tgz)
-
-    # Save it as output
-    RealTimeLogger.get().info("Uploading index of {}".format(
-        graph_filename))
-    index_key = os.path.basename(index_dir_tgz)
-    out_store.write_output_file(index_dir_tgz, index_key)
-    RealTimeLogger.get().info("Index {} uploaded successfully".format(
-        index_key))
-
     #Split fastq files
     return job.addChildJobFn(run_split_fastq, options, index_dir_id, cores=3, memory="4G", disk="2G").rv()
 
@@ -362,7 +151,7 @@ def run_split_fastq(job, options, index_dir_id):
         out_store.write_output_file(filename, filename_key)
         
         RealTimeLogger.get().info("Wrote {} records to {}".format(count, filename))
-
+        
         #Run graph alignment on each fastq chunk
         job.addChildJobFn(run_alignment, options, filename_key, chunk_id, index_dir_id, cores=options.alignment_cores, memory="4G", disk="2G")
     
@@ -666,7 +455,7 @@ def main():
 
     # make the docker runner
     options.drunner = DockerRunner(
-        docker_tool_map = get_chunk_call_docker_tool_map(options))
+        docker_tool_map = get_docker_tool_map(options))
     
     # How long did it take to run the entire pipeline, in seconds?
     run_time_pipeline = None
